Simplify nmax calculation for SphericalExpansion method
The previous version was problematic as it passed back through zero
for sufficiently large gamma. This version uses a linear estimator
which is guaranteed not to have that problem but will use a higher nmax
than before.

This partially addresses #4, althogh a more thorough fix would be
derived from the recurrence formula and would depend on the target
precision.